+++ gradle-jdks/build.gradle
@@ -34,6 +34,7 @@ dependencies {
 
     testImplementation gradleTestKit()
     testImplementation project(':gradle-jdks-settings')
+    testImplementation project(':gradle-jdks-test-common')
     testImplementation 'com.netflix.nebula:nebula-test'
     testImplementation 'org.junit.jupiter:junit-jupiter'
     testImplementation 'org.assertj:assertj-core'
@@ -76,8 +77,7 @@ tasks.withType(JavaCompile).configureEach {
 }
 
 tasks.test {
-    environment(Map.of("PROJECT_VERSION", project.version, "HOME", "/tmp"))
-    systemProperty 'user.home', '/tmp'
+    environment(Map.of("PROJECT_VERSION", project.version))
     systemProperty 'ignoreDeprecations', 'true'
     dependsOn project(':gradle-jdks-settings').tasks.named('build')
 }
