+++ gradle-jdks/src/test/groovy/com/palantir/gradle/jdks/GradleJdkToolchainsIntegrationTest.groovy
@@ -18,20 +18,14 @@ package com.palantir.gradle.jdks
 
 import com.palantir.gradle.jdks.setup.common.CurrentArch
 import com.palantir.gradle.jdks.setup.common.CurrentOs
-import nebula.test.functional.ExecutionResult
 import org.apache.commons.lang3.tuple.Pair
 import spock.lang.TempDir
 
-import java.nio.file.FileVisitResult
-import java.nio.file.Files
 import java.nio.file.Path
-import java.nio.file.SimpleFileVisitor
-import java.nio.file.attribute.BasicFileAttributes
 import java.util.regex.Matcher
 import java.util.regex.Pattern
-import java.util.stream.Stream
 
-class GradleJdkToolchainsIntegrationTest extends GradleJdkIntegrationTest {
+class GradleJdkToolchainsIntegrationTest extends GradleJdkIntegrationSpec {
 
     private static final int JAVA_11_BYTECODE = 55
     private static final int JAVA_17_BYTECODE = 61
@@ -40,71 +34,6 @@ class GradleJdkToolchainsIntegrationTest extends GradleJdkIntegrationTest {
 
     @TempDir
     Path workingDir
-
-
-    def '#gradleVersionNumber: non-existing jdks are installed by the settings plugin'() {
-        gradleVersion = gradleVersionNumber
-        applyJdksPlugins()
-
-        // language=groovy
-        buildFile << """
-            jdks {
-               jdk(17) {
-                  distribution = JDK_17_DISTRO
-                  jdkVersion = JDK_17_VERSION
-               }
-               
-                daemonTarget = '17'
-            }
-        """.replace("JDK_17_DISTRO", quoted(JDK_17.getLeft()))
-                .replace("JDK_17_VERSION", quoted(JDK_17.getRight()))
-                .stripIndent(true)
-
-        when:
-        file('gradle.properties') << 'palantir.jdk.setup.enabled=true'
-        runTasksSuccessfully("generateGradleJdkConfigs")
-
-        then: 'only gradle configuration files are generated, no jdks are installed'
-        String os = CurrentOs.get().uiName()
-        String arch = CurrentArch.get().uiName()
-        Path jdk17LocalPath = projectDir.toPath().resolve("gradle/jdks/17/${os}/${arch}/local-path")
-        String compileJdkFileName = jdk17LocalPath.text.trim()
-        Path installedJdkPath = Path.of(System.getProperty("user.home")).resolve(".gradle/gradle-jdks").resolve(compileJdkFileName).toAbsolutePath()
-        !Files.exists(installedJdkPath)
-
-        when: 'trigger a task'
-        ExecutionResult executionResult = runTasksSuccessfully("javaToolchains")
-
-        then: 'the jdks are installed by the settings plugin'
-        executionResult.standardError.contains("Gradle JDK setup is enabled (palantir.jdk.setup.enabled is true)" +
-                " but some jdks were not installed")
-        executionResult.standardOutput.contains("Auto-detection:     Disabled")
-        executionResult.standardOutput.contains("Auto-download:      Disabled")
-        executionResult.standardOutput.contains("JDK ${SIMPLIFIED_JDK_17_VERSION}")
-        Files.exists(installedJdkPath)
-
-        when: 'if the jdk configured path is changed'
-        jdk17LocalPath.text = "amazon-corretto-another-path\n"
-        ExecutionResult resultAfterJdkChange = runTasksSuccessfully("javaToolchains")
-
-        then:
-        resultAfterJdkChange.standardError.contains("Gradle JDK setup is enabled (palantir.jdk.setup.enabled is true)" +
-                " but some jdks were not installed")
-        Path newInstalledJdkPath = Path.of(System.getProperty("user.home")).resolve(".gradle/gradle-jdks").resolve("amazon-corretto-another-path").toAbsolutePath()
-        Files.exists(newInstalledJdkPath)
-
-        cleanup:
-        Files.walk(installedJdkPath)
-                .sorted(Comparator.reverseOrder())
-                .forEach(Files::delete)
-
-        Files.walk(newInstalledJdkPath)
-                .sorted(Comparator.reverseOrder())
-                .forEach(Files::delete)
-
-        where:
-        gradleVersionNumber << [GRADLE_7_6_VERSION]
-    }
 
     def '#gradleVersionNumber: javaToolchains correctly set-up'() {
         gradleVersion = gradleVersionNumber
@@ -136,9 +65,9 @@ class GradleJdkToolchainsIntegrationTest extends GradleJdkIntegrationTest {
         then: 'the only discovered jdk versions are coming from gradle.properties'
         result.standardOutput.contains("Auto-detection:     Disabled")
         result.standardOutput.contains("Auto-download:      Disabled")
-        result.standardOutput.contains("JDK ${SIMPLIFIED_JDK_11_VERSION}")
-        result.standardOutput.contains("JDK ${SIMPLIFIED_JDK_17_VERSION}")
-        result.standardOutput.contains("JDK ${SIMPLIFIED_JDK_21_VERSION}")
+        result.standardOutput.contains("JDK ${GradleJdkTestUtils.SIMPLIFIED_JDK_11_VERSION}")
+        result.standardOutput.contains("JDK ${GradleJdkTestUtils.SIMPLIFIED_JDK_17_VERSION}")
+        result.standardOutput.contains("JDK ${GradleJdkTestUtils.SIMPLIFIED_JDK_21_VERSION}")
         Matcher matcher = Pattern.compile("Detected by:       (.*)").matcher(result.standardOutput)
         while (matcher.find()) {
             String detectedByPattern = matcher.group(1)
@@ -151,7 +80,7 @@ class GradleJdkToolchainsIntegrationTest extends GradleJdkIntegrationTest {
         then: 'java home is set to out jdk 11 configured version'
         String os = CurrentOs.get().uiName()
         String arch = CurrentArch.get().uiName()
-        String daemonJdkFileName = projectDir.toPath().resolve("gradle/jdks/${DAEMON_MAJOR_VERSION_11}/${os}/${arch}/local-path").text.trim()
+        String daemonJdkFileName = projectDir.toPath().resolve("gradle/jdks/${GradleJdkTestUtils.DAEMON_MAJOR_VERSION_11}/${os}/${arch}/local-path").text.trim()
         Path daemonJvm = workingDir().resolve("gradle-jdks").resolve(daemonJdkFileName).toAbsolutePath()
         gradleHomeOutput.contains("java.home: ${daemonJvm}")
 
@@ -171,7 +100,7 @@ class GradleJdkToolchainsIntegrationTest extends GradleJdkIntegrationTest {
         runOutput.contains("Java home: ${compileJvm}")
 
         where:
-        gradleVersionNumber << [GRADLE_7_6_VERSION, GRADLE_7_6_4_VERSION, GRADLE_8_5_VERSION, GRADLE_8_8_VERSION]
+        gradleVersionNumber << GRADLE_TEST_VERSIONS
     }
 
 
@@ -243,7 +172,7 @@ class GradleJdkToolchainsIntegrationTest extends GradleJdkIntegrationTest {
         readBytecodeVersion(subproject21Class) == Pair.of(0, JAVA_21_BYTECODE)
 
         where:
-        gradleVersionNumber << [GRADLE_7_6_4_VERSION, GRADLE_8_5_VERSION]
+        gradleVersionNumber << GRADLE_TEST_VERSIONS
     }
 
     def '#gradleVersionNumber: fails if the jdk version is not configured'() {
@@ -270,8 +199,8 @@ class GradleJdkToolchainsIntegrationTest extends GradleJdkIntegrationTest {
 
         where:
         gradleVersionNumber  | expectedErrorLines
-        GRADLE_7_6_4_VERSION | ["No compatible toolchains found for request specification: {languageVersion=15, vendor=any, implementation=vendor-specific} (auto-detect false, auto-download false)."]
-        GRADLE_8_5_VERSION   | ["No matching toolchains found for requested specification: {languageVersion=15, vendor=any, implementation=vendor-specific}", "No locally installed toolchains match and toolchain auto-provisioning is not enabled."]
+        GradleJdkTestUtils.GRADLE_7_6_4_VERSION | ["No compatible toolchains found for request specification: {languageVersion=15, vendor=any, implementation=vendor-specific} (auto-detect false, auto-download false)."]
+        GradleJdkTestUtils.GRADLE_8_5_VERSION   | ["No matching toolchains found for requested specification: {languageVersion=15, vendor=any, implementation=vendor-specific}", "No locally installed toolchains match and toolchain auto-provisioning is not enabled."]
     }
 
     def java17PreviewCode = '''
