+++ settings.gradle
@@ -9,4 +9,4 @@ include 'idea-plugin'
 include 'gradle-jdks-settings'
 include 'gradle-jdks-groovy'
 include 'gradle-jdks-enablement'
-
+include 'gradle-jdks-test-common'
